requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-CORS
python-dotenv